Swagger UI url config (#418)
Implement manually `Default` trait for for SwaggerUi config. This fixes 
an issue where custom config or oauth config is provided for `SwaggerUi`
causing SwaggerUI not to render correctly in cases where no url configuration
is provided.

Allow overriding swagger config urls in actix-web, axum and rocket. Prior to
this PR the `SwaggerUi` did override the swagger config urls with urls
provided to it. This made it impossible to give alternative URLs for
Swagger UI to use. This commit will change that.

Co-authored-by: komi1230 <yukke.konan@gmail.com>