Add flex to `utoipa-swagger-ui` build (#845)
This PR changes the fact that Swagger UI was vendored dependency. While this 
makes sure that everyone has a working Swagger UI, it is rather rigid and slow
update cycle for users. This commit updates the Swagger UI to the latest version by
default (5.11.0) and removes the vendored Swagger UI completely. The Swagger UI 
will now be downloaded upon `utoipa-swagger-ui` build while allowing users 
optionally define a version of choice and override files in the Swagger UI. Swagger UI 
can also be defined with a custom path that contains the pre downloaded Swagger UI.

Resolves #844 
---------

Co-authored-by: Juha Kukkonen <juha7kukkonen@gmail.com>